requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
pymongo==4.5.0
python-dotenv==1.0.0
python-socketio==5.9.0
python-engineio==4.7.1
redis==5.0.0
minio==7.1.16
requests==2.31.0
bcrypt==4.0.1
PyJWT==2.8.1
email-validator==2.1.0
